+++ backend-frontend-template/admin/lib/BFT/partials/bft-admin-display-changelog-bft.php
@@ -1,0 +1,56 @@
+<?php
+
+/**
+ * Explaining the next updates
+ *
+ * @link		https://moisesbarrachina.online
+ * @since		1.0.0
+ * @version		1.0.0
+ * @package		BFT
+ * @subpackage	BFT/admin/lib/BFT/partials
+ * @author		Moisés Barrachina Planelles <info@moisesbarrachina.online>
+*/
+
+$this->admin_permission_check();
+?>
+<div class="wrap bft_wrap bft_wrap_ol_ul">
+	<h1><?=$this->esc_html_e($this->admin_title)?></h1>
+	<?=$this->html_tabs(false)?>
+	<h2><?=$this->esc_html_e($title)?></h2>
+	<div>
+		<h3>1.0.0</h3>
+		<p>2023-09-11</p>
+		<p><?=$this->esc_html_e("First stable version of Backend Frontend Template")?></p>
+		<h4>BFT</h4>
+		<ul>
+			<li><?=$this->esc_html_e("Menú system")?></li>
+			<li><?=$this->esc_html_e("Log manipulation")?></li>
+			<li><?=$this->esc_html_e("Error manipulation")?></li>
+			<li><?=$this->esc_html_e("Internationalization")?></li>
+			<li><?=$this->esc_html_e("Frontend system by shortcode")?></li>
+			<li><?=$this->esc_html_e("Internalization into Spanish")?></li>
+		</ul>
+		<h4>BFT Pro</h4>
+		<ul>
+			<li><?=$this->esc_html_e("Menú system with childs")?></li>
+			<li><?=$this->esc_html_e("Settings system")?></li>
+			<li><?=$this->esc_html_e("Inputs system: text, textarea, number, select, select multiple, date, checkbox, checkbox multiple, radio, image, file and input comented")?></li>
+			<li><?=$this->esc_html_e("Example database")?></li>
+			<li><?=$this->esc_html_e("Install and deinstall plugin database")?></li>
+			<li><?=$this->esc_html_e("Automated data manipulation")?></li>
+			<li><?=$this->esc_html_e("Manual data manipulation")?></li>
+			<li><?=$this->esc_html_e("Paginated listing by query")?></li>
+			<li><?=$this->esc_html_e("Direct listing by array")?></li>
+			<li><?=$this->esc_html_e("Manual form")?></li>
+			<li><?=$this->esc_html_e("Log manipulation")?></li>
+			<li><?=$this->esc_html_e("Error manipulation")?></li>
+			<li><?=$this->esc_html_e("Internationalization")?></li>
+			<li><?=$this->esc_html_e("Download system for private files")?></li>
+			<li><?=$this->esc_html_e("Iframe system")?></li>
+			<li><?=$this->esc_html_e("PDF by iframe system")?></li>
+			<li><?=$this->esc_html_e("Frontend system by shortcode")?></li>
+			<li><?=$this->esc_html_e("AJAX frontend system")?></li>
+			<li><?=$this->esc_html_e("Internalization into Spanish")?></li>
+		</ul>
+	</div>
+</div>