+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
-* (3 July 2015). Flattened the folder structure so there is no .org repo parent folder.
-* (4 September 2014). Updating the `README` with Windows symbolic link instructions.
-* (3 September 2014). Updating the `README` to describe how to install the Boilerplate.
-* (1 September 2014). Initial Release.
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.0.0] - 2023-09-04
+
+### Added
+
+- Menú system
+- Log manipulation
+- Error manipulation
+- Internationalization system
+- Frontend system by shortcode
+- Spanish translation from [@moisesbarrachina](https://github.com/moisesbarrachina).